WIP adding UserModel
